Update my_each_spec.rb
I think this test is wrong. I have talked to a couple of learn.co students that have completed this test before, and the second test did not include the requirement for a no "while" loop, in the test. The lab lesson also tells us to use a "while" loop, and the lesson before teaches us how to create our own #each loop using "while".